requirements.txt
appnope==0.1.0
asgiref==3.3.1
autopep8==1.5.4
backcall==0.2.0
decorator==4.4.2
Django==3.1.4
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-rest-marshmallow==4.0.2
djangorestframework==3.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
marshmallow==3.9.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
